% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiasedMatch-package.R
\docType{package}
\name{BiasedMatch-package}
\alias{BiasedMatch}
\alias{BiasedMatch-package}
\title{BiasedMatch: Biased Matching for Two-Urn Sampling Problems}
\description{
Tools for the two-urn biased sampling problems of Puza and Bonfrer (2018) \doi{10.1080/03610926.2017.1300282}. Two overlapping sets of items are sampled without replacement, with items common to both sets given a selection weight favouring their inclusion. The package computes the probability mass function, mean and variance of the number of matching items using the univariate and bivariate Wallenius' non-central hypergeometric distributions, performs maximum likelihood and Bayesian inference for the bias parameter, and provides Metropolis-Hastings algorithms for hierarchical extensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrebonfrer/BiasedMatch}
  \item Report bugs at \url{https://github.com/andrebonfrer/BiasedMatch/issues}
}

}
\author{
\strong{Maintainer}: Andre Bonfrer \email{andre.bonfrer@deakin.edu.au} (\href{https://orcid.org/0000-0002-3061-7103}{ORCID})

Authors:
\itemize{
  \item Andre Bonfrer \email{andre.bonfrer@deakin.edu.au} (\href{https://orcid.org/0000-0002-3061-7103}{ORCID})
  \item Borek Puza
}

}
\keyword{internal}
